+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Unreleased
 Added
 ~~~~~~
 - has_original_id for datasets. Returns true if all pointclouds have original_id
+- PointCloud.from_file now supports timestamp input or "from_file"
 
 Changed
 ~~~~~~